UBUNTU: [Packaging] azure: Add pvpanic kernel modules to linux-modules
BugLink: https://bugs.launchpad.net/bugs/2126659

These modules are currently shipped in linux-modules-extra. They
are useful in virtual environments for allowing the host OS to
detect crashed guests, so it makes sense to ship them in
linux-modules so that they will be included when linux-virtual
is installed.

Signed-off-by: Mehmet Basaran <mehmet.basaran@canonical.com>
Acked-by: Manuel Diewald <manuel.diewald@canonical.com>
Acked-by: Edoardo Canepa <edoardo.canepa@canonical.com>
Signed-off-by: Edoardo Canepa <edoardo.canepa@canonical.com>

(copied from master)

Signed-off-by: John Cabaj <john.cabaj@canonical.com>